requirements.txt
Django==1.3.1
django-celery==2.4.1
psycopg2==2.4.2
python-dateutil==1.5
redis==2.4.10
